Add back ESL remote check and fix index errors
ESL is a status logger for EUS and I didn't notice when we moved from the old PHP version to the new Revolution version that I neglected to include the check for it, so that was missing for months and months.

Also that index fix is that req.url returns a "/" when you go to index but has no file in the url. Express understands it but fs.access let it through so Express was printing an error to the console which looked a bit ugly and was a bug so I fixed it here too.